refactor(projection): relocate web/game/fog into babylon.projection.fog
Program 24 P1 WO-1 (the Hoist, part A): web/game/fog/{reach,ledger,filter,
__init__}.py were already transport-neutral (zero django/babylon.engine
imports) so this is a pure relocation into src/babylon/projection/fog/,
with cross-reference docstrings updated to the new module path.

web/game/fog/{__init__,reach,ledger,filter}.py are now thin re-export
shims delegating to babylon.projection.fog.* so engine_bridge.py's
`from .fog.* import ...` lines and any other legacy game.fog importer
keep working unchanged until the P4 cutover retires web/ entirely.

tests/unit/web/fog/ moves to tests/unit/projection/fog/ with imports
updated to babylon.projection.fog.*; a new tests/unit/web/test_fog_shim.py
pins that each shim re-exports the identical object (not a re-implemented
copy) from its canonical counterpart, so the shim can't rot silently.

Verified: tests/unit/projection/fog + tests/unit/web/{test_fog_shim,
test_import_boundary,test_hex_rollup_fog,test_org_internal_state_fog,
test_solidarity_edge_fog,test_veil}.py all green (135 passed); mypy clean
on the touched files; lint-imports 8/8 contracts kept; zero django imports
under src/babylon/projection.
