+++ tests/unit/projection/fog/__init__.py
@@ -1,0 +1,2 @@
+"""Tests for the Track 1 fog-of-war package (``babylon.projection.fog``,
+relocated from ``web/game/fog/`` by Program 24 P1 WO-1)."""
