--- tests/unit/web/fog/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for the Track 1 fog-of-war package (``web/game/fog/``)."""
