+++ web/game/fog/__init__.py
@@ -1,26 +1,49 @@
-"""Track 1 — The Organizer's Map (2026-07-18, spec-117 §5a).
+"""Legacy re-export shim over :mod:`babylon.projection.fog`.
 
-Fog-of-war lives at the serialization boundary, not the engine: visibility
-is a pure function of ``(graph, intel_ledger)``, computed here in the
-bridge/session layer and never fed back into the simulation. See
-``docs/superpowers/plans/2026-07-18-track1-organizers-map.md``.
+Program 24 P1 WO-1 (the Hoist, part A) relocated this package's real
+implementation to ``babylon.projection.fog`` — it was already
+transport-neutral (zero ``django``/``babylon.engine`` coupling), so the move
+is a pure relocation with no behavior change. This module exists only so
+that ``web/`` code (``engine_bridge.py`` and any callers importing
+``game.fog``/``game.fog.reach``/``game.fog.ledger``/``game.fog.filter``)
+keeps working unchanged until the P4 cutover retires ``web/`` entirely. New
+code should import :mod:`babylon.projection.fog` directly.
 
-- :mod:`game.fog.reach` — the organizing-reach primitive (Task 2): which
-  node ids are within the player org's PRESENCE ∪ SOLIDARITY neighborhood.
-- :mod:`game.fog.ledger` — the intel ledger (Task 3): a session-scoped,
-  append-only, event-sourced record of INVESTIGATE resolutions, and the
-  pure aging function that renders a snapshot exact/approximate/unknown.
-- :mod:`game.fog.filter` — ``apply_fog`` (Task 4): redacts a composer's
-  political fields outside ``reach``/uncovered by the ledger, always
-  returning a new dict. ``engine_bridge.py`` wires this into
-  ``_serialize_territory``/``_state_to_snapshot``/``_build_org_network``
-  and ``get_inspector_node``/``get_inspector_org`` — never the reverse.
-
-Both modules are deliberately free of any ``babylon.engine`` /
-``babylon.models`` / ``babylon.config`` import — the web import-boundary
-test (``tests/unit/web/test_import_boundary.py``) reserves those to
-``engine_bridge.py`` (+ a short allowlist). Coefficients (reach radius,
-aging thresholds) and graph/edge-type identifiers are passed in explicitly
-by callers that already hold them, keeping this package a pure function of
-its inputs — no engine coupling, no hidden state.
+See the canonical package docstring at
+``src/babylon/projection/fog/__init__.py`` for the fog-of-war design
+(organizing reach, the intel ledger, ``apply_fog``).
 """
+
+from __future__ import annotations
+
+from babylon.projection.fog.filter import (
+    ORG_INTERNAL_STATE_FIELDS,
+    ORG_POLITICAL_FIELDS,
+    POLITICAL_FIELDS,
+    apply_fog,
+    political_field_group,
+)
+from babylon.projection.fog.ledger import (
+    IntelEntry,
+    IntelLedger,
+    IntelReading,
+    VisibilityTier,
+    ledger_from_events,
+    read_intel,
+)
+from babylon.projection.fog.reach import organizing_reach
+
+__all__ = [
+    "IntelEntry",
+    "IntelLedger",
+    "IntelReading",
+    "ORG_INTERNAL_STATE_FIELDS",
+    "ORG_POLITICAL_FIELDS",
+    "POLITICAL_FIELDS",
+    "VisibilityTier",
+    "apply_fog",
+    "ledger_from_events",
+    "organizing_reach",
+    "political_field_group",
+    "read_intel",
+]
